cloud-function-src/requirements.txt
google-cloud-bigquery==3.25.0
google-cloud-resource-manager==1.12.5
google-cloud-secret-manager==2.20.2
pandas==2.2.3